Various improvements to building 1-click images.
Bash script now runs with "-e" option set to help mitigate
cascading errors.

Resultant zip file is now named with build numbers of included
images and VMs.

Some basic preparation for Win32 VM when available.  For now, a
placeholder VM is included (based on 1.3-13315 OneClick) to allow
a full image to be built.